chore: move `nargo expand` code to its own crate
