+++ Cargo.toml
@@ -17,6 +17,7 @@ members = [
     "tooling/ast_fuzzer/fuzz",
     "tooling/greybox_fuzzer",
     "tooling/nargo",
+    "tooling/nargo_expand",
     "tooling/nargo_fmt",
     "tooling/nargo_cli",
     "tooling/nargo_toml",
@@ -103,6 +104,7 @@ noirc_printable_type = { path = "compiler/noirc_printable_type" }
 noir_greybox_fuzzer = { path = "tooling/greybox_fuzzer" }
 nargo = { path = "tooling/nargo" }
 nargo_cli = { path = "tooling/nargo_cli" }
+nargo_expand = { path = "tooling/nargo_expand" }
 nargo_fmt = { path = "tooling/nargo_fmt" }
 nargo_toml = { path = "tooling/nargo_toml" }
 noir_lsp = { path = "tooling/lsp" }
