+++ tooling/nargo_cli/src/cli/expand_cmd.rs
@@ -1,26 +1,18 @@
 use clap::Args;
 use fm::FileManager;
-use items::ItemBuilder;
 use nargo::{
     errors::CompileError, insert_all_files_for_workspace_into_file_manager,
     ops::check_crate_and_report_errors, package::Package, parse_all, prepare_package,
     workspace::Workspace,
 };
-use nargo_fmt::ImportsGranularity;
+use nargo_expand::get_expanded_crate;
 use nargo_toml::PackageSelection;
 use noirc_driver::CompileOptions;
-use noirc_frontend::{
-    hir::{ParsedFiles, def_map::ModuleId},
-    parse_program_with_dummy_file,
-};
-use printer::ItemPrinter;
+use noirc_frontend::hir::ParsedFiles;
 
 use crate::errors::CliError;
 
 use super::{LockType, PackageOptions, WorkspaceCommand};
-
-mod items;
-mod printer;
 
 /// Show the result of macro expansion
 #[derive(Debug, Clone, Args, Default)]
@@ -61,12 +53,12 @@ fn expand_package(
     package: &Package,
     compile_options: &CompileOptions,
 ) -> Result<(), CompileError> {
-    let code = get_expanded_package(file_manager, parsed_files, package, compile_options)?;
+    let code = get_expanded_package_or_error(file_manager, parsed_files, package, compile_options)?;
     println!("{code}");
     Ok(())
 }
 
-fn get_expanded_package(
+fn get_expanded_package_or_error(
     file_manager: &FileManager,
     parsed_files: &ParsedFiles,
     package: &Package,
@@ -79,34 +71,5 @@ fn get_expanded_package(
 
     check_crate_and_report_errors(&mut context, crate_id, compile_options)?;
 
-    let root_module_id = context.def_maps[&crate_id].root();
-    let module_id = ModuleId { krate: crate_id, local_id: root_module_id };
-
-    let mut builder = ItemBuilder::new(crate_id, &context.def_interner, &context.def_maps);
-    let item = builder.build_module(module_id);
-
-    let dependencies = &context.crate_graph[context.root_crate_id()].dependencies;
-
-    let mut string = String::new();
-    let mut printer = ItemPrinter::new(
-        crate_id,
-        &context.def_interner,
-        &context.def_maps,
-        dependencies,
-        &mut string,
-    );
-    printer.show_item(item);
-
-    let (parsed_module, errors) = parse_program_with_dummy_file(&string);
-    if errors.is_empty() {
-        let config = nargo_fmt::Config {
-            reorder_imports: true,
-            imports_granularity: ImportsGranularity::Crate,
-            ..Default::default()
-        };
-        Ok(nargo_fmt::format(&string, parsed_module, &config))
-    } else {
-        string.push_str("\n\n// Warning: the generated code has syntax errors");
-        Ok(string)
-    }
+    Ok(get_expanded_crate(&context, crate_id))
 }
