+++ tooling/nargo_expand/Cargo.toml
@@ -1,0 +1,20 @@
+[package]
+name = "nargo_expand"
+version.workspace = true
+authors.workspace = true
+edition.workspace = true
+rust-version.workspace = true
+license.workspace = true
+
+[lints]
+workspace = true
+
+[dependencies]
+fm.workspace = true
+nargo.workspace = true
+nargo_fmt.workspace = true
+noirc_driver.workspace = true
+noirc_errors.workspace = true
+noirc_frontend.workspace = true
+
+[dev-dependencies]
